+++ common/src/main/java/com/walmartlabs/concord/common/ConfigurationUtils.java
@@ -48,7 +48,7 @@ public final class ConfigurationUtils {
         m.put(path[path.length - 1], b);
     }
 
-    public static void delete(Map<String, Object> a, String ... path) {
+    public static void delete(Map<String, Object> a, String... path) {
         Object holder = get(a, path.length - 1, path);
 
         if (holder != null && !(holder instanceof Map)) {
@@ -100,6 +100,29 @@ public final class ConfigurationUtils {
         }
         return result;
     }
+
+    public static Map<String, Object> toNested(String k, Object v) {
+        String[] as = k.split("\\.");
+        if (as.length == 1) {
+            return Collections.singletonMap(k, v);
+        }
+
+        Map<String, Object> m = new HashMap<>();
+        Map<String, Object> root = m;
+
+        for (int i = 0; i < as.length; i++) {
+            if (i + 1 >= as.length) {
+                m.put(as[i], v);
+            } else {
+                Map<String, Object> mm = new HashMap<>();
+                m.put(as[i], mm);
+                m = mm;
+            }
+        }
+
+        return root;
+    }
+
 
     private ConfigurationUtils() {
     }