+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/PayloadParser.java
@@ -1,11 +1,14 @@
 package com.walmartlabs.concord.server.process;
 
+import com.walmartlabs.concord.common.ConfigurationUtils;
 import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.server.MultipartUtils;
 import org.jboss.resteasy.plugins.providers.multipart.InputPart;
 import org.jboss.resteasy.plugins.providers.multipart.MultipartInput;
 import org.sonatype.siesta.ValidationErrorsException;
 
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response.Status;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -19,24 +22,40 @@ import java.util.Map;
 public final class PayloadParser {
 
     public static Payload parse(String instanceId, Path baseDir, MultipartInput input) throws IOException {
-        Map<String, Path> m = new HashMap<>();
+        Map<String, Path> attachments = new HashMap<>();
+        Map<String, Object> req = new HashMap<>();
 
         for (InputPart p : input.getParts()) {
             String name = MultipartUtils.extractName(p);
-
-            Path dst = baseDir.resolve(name);
-            try (InputStream in = p.getBody(InputStream.class, null);
-                 OutputStream out = Files.newOutputStream(dst)) {
-                IOUtils.copy(in, out);
+            if (name == null || name.startsWith("/") || name.contains("..")) {
+                throw new ProcessException(instanceId, "Invalid attachment name: " + name, Status.BAD_REQUEST);
             }
 
-            m.put(name, dst);
+            if (p.getMediaType().isCompatible(MediaType.APPLICATION_OCTET_STREAM_TYPE)) {
+                Path dst = baseDir.resolve(name);
+                try (InputStream in = p.getBody(InputStream.class, null);
+                     OutputStream out = Files.newOutputStream(dst)) {
+                    IOUtils.copy(in, out);
+                }
+
+                attachments.put(name, dst);
+            } else if (p.getMediaType().isCompatible(MediaType.TEXT_PLAIN_TYPE)) {
+                String v = p.getBodyAsString();
+                Map<String, Object> m = ConfigurationUtils.toNested(name, v);
+                req = ConfigurationUtils.deepMerge(req, m);
+            }
         }
 
-        return new Payload(instanceId).putAttachments(m);
+        return new Payload(instanceId)
+                .putHeader(Payload.REQUEST_DATA_MAP, req)
+                .putAttachments(attachments);
     }
 
     public static EntryPoint parseEntryPoint(String entryPoint) {
+        if (entryPoint == null) {
+            return null;
+        }
+
         String[] as = entryPoint.split(":");
         if (as.length < 1) {
             throw new ValidationErrorsException("Invalid entry point format: " + entryPoint);